fix mean

+++ nemo_aligner/algorithms/ppo.py
@@ -450,8 +450,8 @@ class PPOTrainer:
         metrics = {
             "table": table,
             "consumed_samples": prompt_lengths.size(0),
-            "global_response_lengths_mean": response_lengths.mean(),
-            "global_prompt_lengths": prompt_lengths.mean(),
+            "global_response_lengths_mean": response_lengths.float().mean(),
+            "global_prompt_lengths": prompt_lengths.float().mean(),
             "global_rewards": rewards.mean(),
         }
